Modify model config
