+++ src/helper/diagnose-config.ts
@@ -1,14 +1,14 @@
 import { readFileSync } from 'fs'
 import { safeLoad } from 'js-yaml'
-import { Platform } from '../models/config/diagnose'
+import { DiagnosePlatform } from '../models/config/diagnose'
 import { merge } from 'lodash'
 
 export class DiagnoseConfig {
-  config: Platform[];
+  config: DiagnosePlatform[];
 
   constructor(configFolderPath: string) {
-    const docCommon = safeLoad(readFileSync(configFolderPath + '/common.yml', 'utf8')) as Platform[]
-    const docDiagnose = safeLoad(readFileSync(configFolderPath + '/diagnose.yml', 'utf8')) as Platform[]
+    const docCommon = safeLoad(readFileSync(configFolderPath + '/common.yml', 'utf8')) as DiagnosePlatform[]
+    const docDiagnose = safeLoad(readFileSync(configFolderPath + '/diagnose.yml', 'utf8')) as DiagnosePlatform[]
     merge(docCommon, docDiagnose)
     this.config = docCommon
   }
