+++ src/helper/execute-config.ts
@@ -1,15 +1,15 @@
 import { readFileSync } from 'fs'
 import { safeLoad } from 'js-yaml'
-import { Platform } from '../models/config/diagnose'
+import { ExecutePlatform } from '../models/config/execute'
 import { merge } from 'lodash'
 
 export class ExecuteConfig {
-  config: Platform[];
+  config: ExecutePlatform[];
 
   constructor(configFolderPath: string) {
-    const docCommon = safeLoad(readFileSync(configFolderPath + '/common.yml', 'utf8')) as Platform[]
-    const docDiagnose = safeLoad(readFileSync(configFolderPath + '/execute.yml', 'utf8')) as Platform[]
-    merge(docCommon, docDiagnose)
+    const docCommon = safeLoad(readFileSync(configFolderPath + '/common.yml', 'utf8')) as ExecutePlatform[]
+    const docExecute = safeLoad(readFileSync(configFolderPath + '/execute.yml', 'utf8')) as ExecutePlatform[]
+    merge(docCommon, docExecute)
     this.config = docCommon
   }
 }
