+++ src/models/config/common/device.ts
@@ -1,0 +1,4 @@
+export interface Device {
+    name: string;
+    ip: string;
+}
