+++ src/models/config/common/platform.ts
@@ -1,0 +1,6 @@
+import { Device } from './device'
+
+export interface Platform {
+    name: string;
+    devices: Device[];
+}
