+++ src/models/config/execute/device.ts
@@ -1,0 +1,6 @@
+import { ExecuteConfig } from './execute-config'
+import { Device } from '../common/device'
+
+export interface ExecuteDevice extends Device {
+    execute?: ExecuteConfig[];
+}
