+++ src/models/config/execute/index.ts
@@ -1,1 +1,3 @@
 export { ExecuteConfig } from './execute-config'
+export { ExecutePlatform } from './platform'
+export { ExecuteDevice } from './device'
