+++ src/models/config/execute/platform.ts
@@ -1,0 +1,7 @@
+import { ExecuteDevice } from './device'
+import { Platform } from '../common/platform'
+
+export interface ExecutePlatform extends Platform {
+
+    devices: ExecuteDevice[];
+}
